Move TaskController code to TaskListViewModel
This fixes an issue in the architecture, where there were both MVVM and MVC styles.
