+++ App/ViewModels/TaskListViewModel.cs
@@ -1,48 +1,46 @@
 using System.Collections.ObjectModel;
-using System.ComponentModel;
-using System.Runtime.CompilerServices;
 
-public class TaskListViewModel : INotifyPropertyChanged
+public class TaskListViewModel
 {
-    private TaskController taskController = new TaskController();
-    private ObservableCollection<Task> tasks;
+    private Random random;
 
-    public ObservableCollection<Task> Tasks
-    {
-        get { return tasks; }
-        set
-        {
-            tasks = value;
-            OnPropertyChanged();
-        }
-    }
-
-    public event PropertyChangedEventHandler PropertyChanged;
+    public ObservableCollection<Task> Tasks { get; set; }
 
     public TaskListViewModel()
     {
         Tasks = new ObservableCollection<Task>();
     }
 
-    public async void LoadTasks(int userId)
+    public void LoadTasks(int userId)
     {
-        var taskList = await System.Threading.Tasks.Task.Run(() => taskController.GetTaskList(userId));
+        int numTasks = random.Next(1, 11);
 
         Tasks.Clear();
-        foreach (var task in taskList)
+        for (int i = 0; i < numTasks; i++)
         {
+            Task task = GenerateRandomTask();
             Tasks.Add(task);
         }
     }
 
-    public async void DeleteTask(int taskId)
+    private Task GenerateRandomTask()
     {
-        Tasks.Remove(tasks.Single(t => t.Id == taskId));
-        await System.Threading.Tasks.Task.Run(() => taskController.DeleteTask(taskId));
+        int id = random.Next(1000, 10000);
+        string[] titles = { "Buy groceries", "Finish project", "Call mom", "Exercise", "Read a book", "Clean the house" };
+        string title = titles[random.Next(titles.Length)];
+
+        string lorem = "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.";
+
+        string description = lorem.Substring(random.Next(0, 200), random.Next(51, lorem.Length - 201));
+
+        DateTimeOffset created = DateTimeOffset.Now.AddDays(random.Next(1, 3)).AddHours(random.Next(1, 12));
+        DateTimeOffset due = created.AddDays(random.Next(1, 7));
+
+        return new Task(id, title, description, created, due, false);
     }
 
-    protected virtual void OnPropertyChanged([CallerMemberName] string propertyName = null)
+    public void DeleteTask(int taskId)
     {
-        PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
+        Tasks.Remove(Tasks.Single(t => t.Id == taskId));
     }
 }
